Add concerns to set and order by a random seed
